moved many files to their own packages to clean up the root directory
